// web/src/api/models/TypesHAStatusResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TypesConntrackdStatus } from './TypesConntrackdStatus';
import {
    TypesConntrackdStatusFromJSON,
    TypesConntrackdStatusFromJSONTyped,
    TypesConntrackdStatusToJSON,
    TypesConntrackdStatusToJSONTyped,
} from './TypesConntrackdStatus';
import type { TypesVRRPInstanceStatus } from './TypesVRRPInstanceStatus';
import {
    TypesVRRPInstanceStatusFromJSON,
    TypesVRRPInstanceStatusFromJSONTyped,
    TypesVRRPInstanceStatusToJSON,
    TypesVRRPInstanceStatusToJSONTyped,
} from './TypesVRRPInstanceStatus';

/**
 * 
 * @export
 * @interface TypesHAStatusResponse
 */
export interface TypesHAStatusResponse {
    /**
     * 
     * @type {TypesConntrackdStatus}
     * @memberof TypesHAStatusResponse
     */
    conntrackd?: TypesConntrackdStatus;
    /**
     * 
     * @type {Array<TypesVRRPInstanceStatus>}
     * @memberof TypesHAStatusResponse
     */
    vrrp?: Array<TypesVRRPInstanceStatus>;
}

/**
 * Check if a given object implements the TypesHAStatusResponse interface.
 */
export function instanceOfTypesHAStatusResponse(value: object): value is TypesHAStatusResponse {
    return true;
}

export function TypesHAStatusResponseFromJSON(json: any): TypesHAStatusResponse {
    return TypesHAStatusResponseFromJSONTyped(json, false);
}

export function TypesHAStatusResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesHAStatusResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'conntrackd': json['conntrackd'] == null ? undefined : TypesConntrackdStatusFromJSON(json['conntrackd']),
        'vrrp': json['vrrp'] == null ? undefined : ((json['vrrp'] as Array<any>).map(TypesVRRPInstanceStatusFromJSON)),
    };
}

export function TypesHAStatusResponseToJSON(json: any): TypesHAStatusResponse {
    return TypesHAStatusResponseToJSONTyped(json, false);
}

export function TypesHAStatusResponseToJSONTyped(value?: TypesHAStatusResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'conntrackd': TypesConntrackdStatusToJSON(value['conntrackd']),
        'vrrp': value['vrrp'] == null ? undefined : ((value['vrrp'] as Array<any>).map(TypesVRRPInstanceStatusToJSON)),
    };
}