// web/src/api/models/TypesStatsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TypesBGPStats } from './TypesBGPStats';
import {
    TypesBGPStatsFromJSON,
    TypesBGPStatsFromJSONTyped,
    TypesBGPStatsToJSON,
    TypesBGPStatsToJSONTyped,
} from './TypesBGPStats';
import type { TypesIfaceStats } from './TypesIfaceStats';
import {
    TypesIfaceStatsFromJSON,
    TypesIfaceStatsFromJSONTyped,
    TypesIfaceStatsToJSON,
    TypesIfaceStatsToJSONTyped,
} from './TypesIfaceStats';
import type { TypesOSPFStats } from './TypesOSPFStats';
import {
    TypesOSPFStatsFromJSON,
    TypesOSPFStatsFromJSONTyped,
    TypesOSPFStatsToJSON,
    TypesOSPFStatsToJSONTyped,
} from './TypesOSPFStats';
import type { TypesSystemStats } from './TypesSystemStats';
import {
    TypesSystemStatsFromJSON,
    TypesSystemStatsFromJSONTyped,
    TypesSystemStatsToJSON,
    TypesSystemStatsToJSONTyped,
} from './TypesSystemStats';

/**
 * 
 * @export
 * @interface TypesStatsResponse
 */
export interface TypesStatsResponse {
    /**
     * 
     * @type {TypesBGPStats}
     * @memberof TypesStatsResponse
     */
    bgp?: TypesBGPStats;
    /**
     * 
     * @type {{ [key: string]: TypesIfaceStats; }}
     * @memberof TypesStatsResponse
     */
    interfaces?: { [key: string]: TypesIfaceStats; };
    /**
     * 
     * @type {TypesOSPFStats}
     * @memberof TypesStatsResponse
     */
    ospf?: TypesOSPFStats;
    /**
     * 
     * @type {TypesSystemStats}
     * @memberof TypesStatsResponse
     */
    system?: TypesSystemStats;
}

/**
 * Check if a given object implements the TypesStatsResponse interface.
 */
export function instanceOfTypesStatsResponse(value: object): value is TypesStatsResponse {
    return true;
}

export function TypesStatsResponseFromJSON(json: any): TypesStatsResponse {
    return TypesStatsResponseFromJSONTyped(json, false);
}

export function TypesStatsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesStatsResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'bgp': json['bgp'] == null ? undefined : TypesBGPStatsFromJSON(json['bgp']),
        'interfaces': json['interfaces'] == null ? undefined : (mapValues(json['interfaces'], TypesIfaceStatsFromJSON)),
        'ospf': json['ospf'] == null ? undefined : TypesOSPFStatsFromJSON(json['ospf']),
        'system': json['system'] == null ? undefined : TypesSystemStatsFromJSON(json['system']),
    };
}

export function TypesStatsResponseToJSON(json: any): TypesStatsResponse {
    return TypesStatsResponseToJSONTyped(json, false);
}

export function TypesStatsResponseToJSONTyped(value?: TypesStatsResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'bgp': TypesBGPStatsToJSON(value['bgp']),
        'interfaces': value['interfaces'] == null ? undefined : (mapValues(value['interfaces'], TypesIfaceStatsToJSON)),
        'ospf': TypesOSPFStatsToJSON(value['ospf']),
        'system': TypesSystemStatsToJSON(value['system']),
    };
}