// web/src/api/models/TypesSystemHistoryPoint.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesSystemHistoryPoint
 */
export interface TypesSystemHistoryPoint {
    /**
     * 
     * @type {number}
     * @memberof TypesSystemHistoryPoint
     */
    cpu_pct: number;
    /**
     * 
     * @type {number}
     * @memberof TypesSystemHistoryPoint
     */
    load1: number;
    /**
     * 
     * @type {number}
     * @memberof TypesSystemHistoryPoint
     */
    mem_total: number;
    /**
     * 
     * @type {number}
     * @memberof TypesSystemHistoryPoint
     */
    mem_used: number;
    /**
     * 
     * @type {number}
     * @memberof TypesSystemHistoryPoint
     */
    ts: number;
}

/**
 * Check if a given object implements the TypesSystemHistoryPoint interface.
 */
export function instanceOfTypesSystemHistoryPoint(value: object): value is TypesSystemHistoryPoint {
    if (!('cpu_pct' in value) || value['cpu_pct'] === undefined) return false;
    if (!('load1' in value) || value['load1'] === undefined) return false;
    if (!('mem_total' in value) || value['mem_total'] === undefined) return false;
    if (!('mem_used' in value) || value['mem_used'] === undefined) return false;
    if (!('ts' in value) || value['ts'] === undefined) return false;
    return true;
}

export function TypesSystemHistoryPointFromJSON(json: any): TypesSystemHistoryPoint {
    return TypesSystemHistoryPointFromJSONTyped(json, false);
}

export function TypesSystemHistoryPointFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesSystemHistoryPoint {
    if (json == null) {
        return json;
    }
    return {
        
        'cpu_pct': json['cpu_pct'],
        'load1': json['load1'],
        'mem_total': json['mem_total'],
        'mem_used': json['mem_used'],
        'ts': json['ts'],
    };
}

export function TypesSystemHistoryPointToJSON(json: any): TypesSystemHistoryPoint {
    return TypesSystemHistoryPointToJSONTyped(json, false);
}

export function TypesSystemHistoryPointToJSONTyped(value?: TypesSystemHistoryPoint | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'cpu_pct': value['cpu_pct'],
        'load1': value['load1'],
        'mem_total': value['mem_total'],
        'mem_used': value['mem_used'],
        'ts': value['ts'],
    };
}