// web/src/api/models/TypesNeighborStat.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesNeighborStat
 */
export interface TypesNeighborStat {
    /**
     * 
     * @type {string}
     * @memberof TypesNeighborStat
     */
    dev: string;
    /**
     * 
     * @type {string}
     * @memberof TypesNeighborStat
     */
    family: string;
    /**
     * 
     * @type {string}
     * @memberof TypesNeighborStat
     */
    hostname?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesNeighborStat
     */
    ip: string;
    /**
     * 
     * @type {string}
     * @memberof TypesNeighborStat
     */
    mac?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesNeighborStat
     */
    state: string;
}

/**
 * Check if a given object implements the TypesNeighborStat interface.
 */
export function instanceOfTypesNeighborStat(value: object): value is TypesNeighborStat {
    if (!('dev' in value) || value['dev'] === undefined) return false;
    if (!('family' in value) || value['family'] === undefined) return false;
    if (!('ip' in value) || value['ip'] === undefined) return false;
    if (!('state' in value) || value['state'] === undefined) return false;
    return true;
}

export function TypesNeighborStatFromJSON(json: any): TypesNeighborStat {
    return TypesNeighborStatFromJSONTyped(json, false);
}

export function TypesNeighborStatFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesNeighborStat {
    if (json == null) {
        return json;
    }
    return {
        
        'dev': json['dev'],
        'family': json['family'],
        'hostname': json['hostname'] == null ? undefined : json['hostname'],
        'ip': json['ip'],
        'mac': json['mac'] == null ? undefined : json['mac'],
        'state': json['state'],
    };
}

export function TypesNeighborStatToJSON(json: any): TypesNeighborStat {
    return TypesNeighborStatToJSONTyped(json, false);
}

export function TypesNeighborStatToJSONTyped(value?: TypesNeighborStat | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'dev': value['dev'],
        'family': value['family'],
        'hostname': value['hostname'],
        'ip': value['ip'],
        'mac': value['mac'],
        'state': value['state'],
    };
}